Improve [CI/CD] [Docker Builder] [Private Registry] Support TAG argument
- [+] feat(workflows): add build-args to private-registry.yaml
- [+] feat(Dockerfile): add TAG argument to go build command
- [+] docs(README.md): update project structure
- [+] feat(run.go): add build tag for non-immutable
- [+] feat(run_immutable.go): add Config struct
- [+] refactor(routes.go): remove RegisterRoutes function
- [+] feat(routes_non_immutable.go): add RegisterRoutes for non-immutable build
