+++ .github/workflows/private-registry.yaml
@@ -61,7 +61,9 @@ jobs:
           # Note: By default, the architecture is linux/amd64,linux/arm64 because Sonic JSON requires amd64,arm64 (see https://github.com/bytedance/sonic).
           # Tip: If your cloud provider supports multi-architecture nodes for your cluster, it will be great for smooth sailing ⛵ ☸ (HPA + Cluster Autoscaling/Autopilot).
           platforms: linux/amd64,linux/arm64
-          # Note: Don't forget to modify "example.com"
+          # Note: Don't forget to modify "example.com", also "build-args." This is part of Variables Management, which is also supported in Gitea.
           tags: |
             example.com/${{ secrets.REPO_NAME }}/${{ secrets.IMAGE_NAME }}:${{ github.ref_name }}
             example.com/${{ secrets.REPO_NAME }}/${{ secrets.IMAGE_NAME }}:latest
+          build-args: |
+            TAG=${{ vars.TAG }}
