+++ backend/cmd/server/run.go
@@ -3,6 +3,9 @@
 // By accessing or using this software, you agree to be bound by the terms
 // of the License Agreement, which you can find at LICENSE files.
 
+//go:build !immutable
+// +build !immutable
+
 package main
 
 import (
